sensor_manager/requirements.txt
kafka-python
pymongo
flask
requests
faker